Update query helper to reflect post-2.3.0 changes
This change modifies the BigQueryQueryHelper code to integrate new
changes to the query code, adding support for query priority and
location.
